requirements.txt
langchain-anthropic==0.3.10
python-dotenv==1.0.0
